Import DRI-DroboFSLinuxKernel.tgz

+++ kernel/include/asm-arm/unaligned.h
@@ -40,24 +40,24 @@ extern int __bug_unaligned_x(const void *ptr);
  */
 
 #define __get_unaligned_2_le(__p)					\
-	(__p[0] | __p[1] << 8)
+	(unsigned int)(__p[0] | __p[1] << 8)
 
 #define __get_unaligned_2_be(__p)					\
-	(__p[0] << 8 | __p[1])
+	(unsigned int)(__p[0] << 8 | __p[1])
 
 #define __get_unaligned_4_le(__p)					\
-	(__p[0] | __p[1] << 8 | __p[2] << 16 | __p[3] << 24)
+	(unsigned int)(__p[0] | __p[1] << 8 | __p[2] << 16 | __p[3] << 24)
 
 #define __get_unaligned_4_be(__p)					\
-	(__p[0] << 24 | __p[1] << 16 | __p[2] << 8 | __p[3])
+	(unsigned int)(__p[0] << 24 | __p[1] << 16 | __p[2] << 8 | __p[3])
 
 #define __get_unaligned_8_le(__p)					\
-	((unsigned long long)__get_unaligned_4_le((__p+4)) << 32 |	\
-		__get_unaligned_4_le(__p))
+	(((unsigned long long)(unsigned long)__get_unaligned_4_le((__p+4))) << 32 |	\
+		(unsigned long)__get_unaligned_4_le(__p))
 
 #define __get_unaligned_8_be(__p)					\
-	((unsigned long long)__get_unaligned_4_be(__p) << 32 |		\
-		__get_unaligned_4_be((__p+4)))
+	(((unsigned long long)(unsigned long)__get_unaligned_4_be(__p)) << 32 |		\
+		(unsigned long)__get_unaligned_4_be((__p+4)))
 
 #define __get_unaligned_le(ptr)						\
 	({								\
